Backend/requirements.txt
fastapi
uvicorn
mistralai
pandas
python-dotenv
pydantic